requirements.txt
# Core libraries
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0

# Data handling
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.1
datasets == 3.0.1

# Image processing & augmentations
Pillow==10.4.0
albumentations==1.4.7
opencv-python==4.10.0.84

# Visualization
matplotlib==3.9.2
seaborn==0.13.2
tqdm==4.66.5

# Optional for embeddings visualization
umap-learn==0.5.6
scipy==1.13.1
